requirements-company.txt
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.1.2
Flask==0.12.2
gevent==1.2.2
greenlet==0.4.12
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.9.6
Keras==2.0.8
Markdown==2.6.9
MarkupSafe==1.0
numpy==1.13.3
pandas==0.21.0
parso==0.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.4.0
psycopg2==2.7.3.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.14
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
